refactor(docfx): split page-related functions
I plan to add additional functions related to pages, and the one source
file was getting too big already.
