+++ docfx/CMakeLists.txt
@@ -29,7 +29,8 @@ include(EnableWerror)
 
 add_library(
     docfx # cmake-format: sort
-    doxygen2markdown.cc doxygen2markdown.h parse_arguments.cc parse_arguments.h)
+    doxygen2markdown.cc doxygen2markdown.h doxygen_pages.cc doxygen_pages.h
+    parse_arguments.cc parse_arguments.h)
 google_cloud_cpp_add_common_options(docfx)
 target_include_directories(docfx PUBLIC "${PROJECT_SOURCE_DIR}")
 target_compile_features(docfx PUBLIC cxx_std_17)
@@ -50,8 +51,9 @@ if (NOT BUILD_TESTING)
     return()
 endif ()
 
-set(unit_tests # cmake-format: sort
-               "doxygen2markdown_test.cc" "parse_arguments_test.cc")
+set(unit_tests
+    # cmake-format: sort
+    doxygen2markdown_test.cc doxygen_pages_test.cc parse_arguments_test.cc)
 
 # Export the list of unit tests to a .bzl file so we do not need to maintain the
 # list in two places.
