+++ docfx/docfx.bzl
@@ -18,10 +18,12 @@
 
 docfx_hdrs = [
     "doxygen2markdown.h",
+    "doxygen_pages.h",
     "parse_arguments.h",
 ]
 
 docfx_srcs = [
     "doxygen2markdown.cc",
+    "doxygen_pages.cc",
     "parse_arguments.cc",
 ]
